requirements.txt
streamlit==1.28.1
requests==2.31.0
pandas==2.1.3
plotly==5.17.0
openai==1.3.7
python-dotenv==1.0.0
rdkit==2023.9.1
pubchempy==1.0.4
beautifulsoup4==4.12.2
lxml==4.9.3
